Move ingestors; Add SER and TIF ingestor, add DM test
