--- xicam/NCEM/SERPlugin.yapsy-plugin
@@ -1,9 +1,0 @@
-[Core]
-Name = SERPlugin
-Module = SERPlugin.py
-
-[Documentation]
-Author = Ronald J. Pandolfi
-Version = 0.1.0
-Website = http://lotsofplugins.com
-Description = My first plugin
